+++ include/font_def.hpp
@@ -5,14 +5,14 @@
 #include <string>
 
 struct Text {
-  SDL_Texture* tex;
-  SDL_Surface* surf;
-  SDL_Rect     rect;
-  int          width;
-  int          height;
-  int          id;
-  bool         is_valid;
-  std::string  name;
+  SDL_Texture *tex;
+  SDL_Surface *surf;
+  SDL_Rect rect;
+  int width;
+  int height;
+  size_t id;
+  bool is_valid;
+  std::string name;
 };
 
 typedef Text Text;
